refactored and removed old exoplayer implementation
