+++ feature/player/src/main/java/com/arcticoss/nextplayer/feature/player/composables/MediaGestures.kt
@@ -1,6 +1,5 @@
-package com.arcticoss.nextplayer.feature.player.compose
+package com.arcticoss.nextplayer.feature.player.composables
 
-import android.util.Log
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -8,9 +7,9 @@ import androidx.compose.foundation.layout.systemGesturesPadding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.pointer.pointerInput
-import com.arcticoss.nextplayer.feature.player.presentation.ControllerState
 import com.arcticoss.nextplayer.feature.player.presentation.ControllerVisibility
 import com.arcticoss.nextplayer.feature.player.presentation.MediaState
+import com.arcticoss.nextplayer.feature.player.state.ControllerState
 
 
 private const val TAG = "MediaGestures"
