+++ feature/player/src/main/java/com/arcticoss/nextplayer/feature/player/composables/VerticalSwipeMediaControls.kt
@@ -1,4 +1,4 @@
-package com.arcticoss.nextplayer.feature.player.presentation.composables
+package com.arcticoss.nextplayer.feature.player.composables
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.LightMode
@@ -6,6 +6,7 @@ import androidx.compose.material.icons.rounded.VolumeMute
 import androidx.compose.material.icons.rounded.VolumeUp
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import com.arcticoss.nextplayer.feature.player.presentation.composables.AdjustmentBar
 
 @Composable
 fun BrightnessAdjustmentBar(
